chore(ci): stop firestore query-coverage counts from flapping the ratchet
Keep failing on new unregistered/unsupported shapes, stale debt entries,
and registered-count or percentage decreases. Treat live counts above the
committed file as a pass with a regenerate notice so two merges no longer
race the baseline-equality check.
